requirements.txt
absl-py==2.1.0
accelerate==1.6.0
aider-chat==0.73.0
aider-install==0.1.3
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anthropic==0.34.2
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==25.1.0
babel==2.16.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.12.3
black==23.3.0
bleach==6.1.0
blinker==1.9.0
bokeh==3.6.3
cachetools==5.5.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
colorcet==3.1.0
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.3.1
cycler==0.12.1
dataclasses-json==0.5.7
datasets==2.21.0
datashader==0.17.0
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
diff-match-patch==20241021
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
executing==2.0.1
faiss-cpu==1.9.0.post1
fastapi==0.115.12
fastjsonschema==2.20.0
ffmpy==0.5.0
filelock==3.17.0
flake8==7.1.1
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==25.1.24
fonttools==4.55.8
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.12.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
-e git+https://github.com/gpt-engineer-org/gpt-engineer.git@4fea4f03976e76a7723306e5acecca57cf68554c#egg=gpt_engineer
gradio==5.29.1
gradio_client==1.10.1
greenlet==3.1.1
grep-ast==0.4.1
groovy==0.1.2
groq==0.13.0
grpcio==1.70.0
h11==0.14.0
h5py==3.12.1
holoviews==1.20.1
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.31.2
hvplot==0.11.2
idna==3.10
imbalanced-learn==0.13.0
importlib_metadata==7.2.1
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.27.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kaggle==1.6.17
kagglehub==0.3.10
keras==3.8.0
kiwisolver==1.4.8
langchain==0.3.9
langchain-anthropic==0.1.23
langchain-community==0.3.9
langchain-core==0.3.21
langchain-groq==0.2.1
langchain-huggingface==0.1.2
langchain-openai==0.1.25
langchain-text-splitters==0.3.2
langsmith==0.1.129
libclang==18.1.1
linkify-it-py==2.0.3
litellm==1.59.8
llvmlite==0.44.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.22.0
marshmallow-enum==1.5.1
matplotlib==3.10.0
matplotlib-inline==0.1.7
mccabe==0.7.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
mistune==3.0.2
mixpanel==4.10.1
ml-dtypes==0.4.1
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
multipledispatch==1.0.0
multiprocess==0.70.16
mypy-extensions==1.0.0
namex==0.0.8
narwhals==1.35.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
notebook==7.2.2
notebook_shim==0.2.4
numba==0.61.0
numpy==1.26.4
openai==1.60.2
opt_einsum==3.4.0
optree==0.14.0
orjson==3.10.7
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.2.2
pandocfilters==1.5.1
panel==1.6.1
param==2.2.0
parso==0.8.4
passlib==1.7.4
pathspec==0.12.1
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
posthog==3.11.0
prometheus_client==0.20.0
prompt_toolkit==3.0.50
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.4.8
pycodestyle==2.12.1
pycparser==2.22
pyct==0.5.0
pydantic==2.10.6
pydantic-settings==2.6.1
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
pyflakes==3.2.0
Pygments==2.19.1
pymongo==4.10.1
pypandoc==1.15
pyparsing==3.2.1
pyperclip==1.9.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-http-client==3.3.7
python-jose==3.4.0
python-json-logger==2.0.7
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2024.1
pyviz_comms==3.0.4
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
rsa==4.9.1
rudder-sdk-python==2.1.4
ruff==0.11.10
safehttpx==0.1.6
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
selenium==4.30.0
semantic-version==2.10.0
Send2Trash==1.8.3
sendgrid==6.11.0
sentence-transformers==3.3.1
sentencepiece==0.2.0
setuptools==75.8.0
shap==0.47.0
shellingham==1.5.4
six==1.17.0
sklearn-compat==0.1.3
slicer==0.0.8
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.1
soundfile==0.13.1
soupsieve==2.6
SQLAlchemy==2.0.35
stack-data==0.6.3
starkbank-ecdsa==2.2.0
starlette==0.46.1
streamlit==1.44.1
sympy==1.13.1
tabulate==0.9.0
tenacity==8.5.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-hub==0.16.1
tensorflow_intel==2.18.0
termcolor==2.3.0
terminado==0.18.1
text-unidecode==1.3
textblob==0.18.0.post0
tf_keras==2.18.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tinycss2==1.3.0
tokenizers==0.20.3
toml==0.10.2
tomlkit==0.12.5
toolz==1.0.0
torch==2.6.0+cu118
torchaudio==2.6.0+cu118
torchvision==0.21.0
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.46.3
tree-sitter==0.21.3
tree-sitter-languages==1.10.2
trio==0.29.0
trio-websocket==0.12.2
typer==0.12.5
types-python-dateutil==2.9.0.20240821
typing-inspect==0.9.0
typing_extensions==4.13.0
tzdata==2024.1
uc-micro-py==1.0.3
uri-template==1.3.0
urllib3==2.3.0
uv==0.5.27
uvicorn==0.34.0
watchdog==6.0.0
watchfiles==1.0.4
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.13
wrapt==1.17.2
wsproto==1.2.0
xarray==2025.1.2
xgboost==2.1.4
xxhash==3.5.0
xyzservices==2025.1.0
yarl==1.18.3
zipp==3.21.0
